Product model commited to RSG- 55 feature

+++ hibernateexample/src/main/java/com/aa/hibernate/Delete_Record.java
@@ -17,7 +17,7 @@ public class Delete_Record {
 		
 		Session session = factory.openSession();
 		
-		Object obj = session.load(User_Detail.class, new Integer(5));
+		Object obj = session.load(User_Detail.class, new Integer(6));
 		
 		 User_Detail user= (User_Detail)obj;
